Introduce utils class
