+++ custom/src/index.js
@@ -1,12 +1,11 @@
 const interactions = require("./interactions");
+const Utils = require("./utils");
 
 async function doInteraction(interaction) {
   const start = Date.now();
   await interaction();
   console.log(`${interaction.name} took ${Date.now() - start} ms`);
 }
-
-const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
 
 async function scheduleInteractions() {
   while (true) {
@@ -16,7 +15,7 @@ async function scheduleInteractions() {
     );
 
     // wait some time, between half a second to five seconds between interactions
-    await delay(500 + Math.floor(Math.random() * 4500));
+    await Utils.delay(500 + Math.floor(Math.random() * 4500));
   }
 }
